pyproject.toml
[tool.poetry]
name = "evergreen_scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "AGPL-3.0"

packages = [
  { include = "cli", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.8"
evergreen-py = "^3.10.6"

[tool.poetry.dev-dependencies]
ruff = "^0.9.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
analyze-patch = "cli.analyze_patch:main"
viewless-suites = "cli.viewless_suites:main"
